Renamed EditTask view to DetailTask. Refined the UI of this page.
